+++ lib/features/starter_packs/presentation/starter_pack_detail_screen.dart
@@ -42,7 +42,7 @@ class _StarterPackDetailView extends StatelessWidget {
           if (context.canPop()) {
             context.pop();
           } else {
-            context.go('/profile');
+            context.go('/profile/me');
           }
           return;
         }